hu 17 eliminar eventos con musicos asociados

+++ CastillaCanoAlvaro-Proyecto1/src/main/java/com/salesianostriana/dam/castillacanoalvaroproyecto1/controlador/EventoControlador.java
@@ -1,6 +1,5 @@
 package com.salesianostriana.dam.castillacanoalvaroproyecto1.controlador;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
@@ -14,11 +13,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.salesianostriana.dam.castillacanoalvaroproyecto1.modelo.Concierto;
-import com.salesianostriana.dam.castillacanoalvaroproyecto1.modelo.Evento;
 import com.salesianostriana.dam.castillacanoalvaroproyecto1.modelo.Procesion;
+import com.salesianostriana.dam.castillacanoalvaroproyecto1.servicio.AsisteServicio;
 import com.salesianostriana.dam.castillacanoalvaroproyecto1.servicio.BusServicio;
 import com.salesianostriana.dam.castillacanoalvaroproyecto1.servicio.ConciertoServicio;
 import com.salesianostriana.dam.castillacanoalvaroproyecto1.servicio.EventoServicio;
+import com.salesianostriana.dam.castillacanoalvaroproyecto1.servicio.MusicoServicio;
 import com.salesianostriana.dam.castillacanoalvaroproyecto1.servicio.ProcesionServicio;
 
 @Controller
@@ -36,6 +36,12 @@ public class EventoControlador {
 
 	@Autowired
 	private ConciertoServicio servicioConcer;
+
+	@Autowired
+	private AsisteServicio asisteServicio;
+
+	@Autowired
+	private MusicoServicio musicoServicio;
 
 	@GetMapping("/listadoEvento")
 	public String lista(Model model) {
@@ -85,7 +91,13 @@ public class EventoControlador {
 
 	@GetMapping("/eliminar/procesion/{id}")
 	public String eliminarProcesion(@PathVariable("id") long id) {
-		servicioProce.deleteById(id);
+		Optional<Procesion> procesion = servicioProce.findBYId(id);
+
+		if (procesion.isPresent()) {
+			servicio.eleminarAsistentesEventos(procesion.get());
+
+			servicioProce.deleteById(id);
+		}
 
 		return "redirect:/admin/evento/listadoEvento";
 	}
@@ -127,11 +139,14 @@ public class EventoControlador {
 
 	@GetMapping("/eliminar/concierto/{id}")
 	public String eliminarConcierto(@PathVariable("id") long id) {
-		servicioConcer.deleteById(id);
+		Optional<Concierto> concierto = servicioConcer.findBYId(id);
 
+		if (concierto.isPresent()) {
+			servicio.eleminarAsistentesEventos(concierto.get());
+
+			servicioConcer.deleteById(id);
+		}
 		return "redirect:/admin/evento/listadoEvento";
 	}
-
-
 
 }
